+++ lib/src/cubit_with_effects.dart
@@ -19,7 +19,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 ///
 /// {@endtemplate}
 abstract class CubitWithEffects<State, Effect> extends Cubit<State>
-    with Effects<Effect> {
+    with Effects<Effect>
+    implements BlocEffectsSource<State, Effect> {
   /// {@macro cubit}
   CubitWithEffects(super.initialState);
 }